+++ Departments/Compliance/__init__.py
@@ -1,0 +1,4 @@
+"""Compliance Department"""
+from .compliance_department import ComplianceDepartment
+
+__all__ = ['ComplianceDepartment']